backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.0",
    "django>=5.2.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "whitenoise>=6.9.0",
]
